Convert all instances of 64bit logic to use memcpy
Platforms (such as particle) don't allow pointer-casting for 64 bit memory
space on non-aligned processor addresses. Meaning buf+4 is fine, but buf+1/2/3 is not.

This moves that logic to be platform specific.

This change will also provide better logging for 64 bit numbers (convert to hex), which
seemed to sometimes cause issues as well